HOL-Light: Run proofs on mlkem-native assembly
This commit uses the `simpasm` script to auto-generate the
input assembly for the HOL-Light proofs from the assembly that's
shipped with mlkem-native.

A minor adjustment is made to `simpasm` along the way, changing
the meaning of `-p` from preserving the header/footer to preserving
the preprocessor directives in the header. That is, the header will
_always_ be preserved, and if `-p` is not set, then preprocessor
directives will be dropped, making the resulting assembly file
standalone. We always preserve to header to ensure that licensing
notices are kept.

The `autogen` script is adjusted to ensure that the assembly sources
in `proofs/hol_light/arm/mlkem` are in sync with `mlkem/*`. Note that
`simpasm` in particular ensures that simplified assembly has the same
byte code as the original one, so the HOL-Light proofs apply to the
assembly shipped in mlkem-native. Previous warning notices in READMEs
are therefore removed.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>