Only run constant-time tests using patched Valgrind
We were running the Valgrind tests both with stock Valgrind, and
a patched version to also detect variable-latency instructions.
The latter can detect anything the former can and since both are working
fine in CI, we can simply drop running the former in CI.

This may cause problems in the future when updating Valgrind in case
the patch no longer applies. But we can still switch back in that case.
I've left a note.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>