Monobuild: Build library file as intermediate
This commit modifies the examples `monolithic_build` and
`monolithic_build_multilevel` to create `libmlkemXXX.a` build artifacts
as intermediate results, before linking against the application source.

Those intermediates are good targets for future code-size measurements,
because in a monobuild the compiler has visibility into what is public
and what is internal.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>