autogen: Don't #undef macros from K-generic headers
Previously, the monobuild file could be configured to be traceless in the
sense that it #undef's all macros defined in any header in the mlkem-native
source tree -- including K-generic headers.

The primary purpose of this `#undef`ing is to facilitate SCU builds with
multiple security levels, in which case K-specific headers will be included
multiple times. `#undef`ing the K-generic shared headers, in turn, must in
fact not be done so long as further instances of the monobuild CU are to be
included, while the very last inclusion, it is unneeded. The cost of the
`#undef`ing is a inpalletable mlkem_native_monobuild.c file.

This commit simplifies mlkem_native_monobuild.c by never `#undef`ing macros
from K-generich header files. Those will therefore 'escape' the MLKEM code
region in an SCU build, but no problems are prevented by this which could
not occur otherwise: For example, if the SCU build includes other libraries
or components which clash in terms of the defined macros, that could already
surface if the conflicting file is included before the mlkem-native CUs, and
starting to rely on inclusion order is brittle -- if there is a conflict, naming
needs to be changed.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>