Examples: Add KAT-check
Previously, all examples would only run a key consistency check,
but not compare against known good values, which can hide issues.

This commit adds a KAT check to all examples. The hardcoded shared
secret is the one generated using the `notrandombytes` RNG in its
initial configuration. To reproduce those values in multilevel
examples, the RNG needs to be reset between runs, for which an
API `randombytes_reset()` is added.

This change highlights an issue with the example
monolithic_build_multilevel_native, which currently fails. This
will be addressed in the next commit.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>