Native Example: Add appropriate CFLAGS
The examples `monolithic_build_multilevel_native` and
`multilevel_build_native` exercise native code and hence
need appropriate CFLAGS on x86 to enable AVX2. On AArch64,
no flags are needed, but FORCE_AARCH64 is still useful to
be passed to ensure that the architecture is correctly
determined.

This commit inlines the contents of test/mk/auto.mk for
system-specific processor flags into the Makefiles of
said examples.

The commit also fixes some multilevel build guards in
the multilevel_build_native example.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>