Autogen: Don't undef'ine macros from header files
A recent commit modified `autogen` so that every .c file includes
a footer of `#undef`s for all macros defined in the file.

This is problematic with `#define`s clauses which are not original
to the .c file, but are repetitions meant as consistency checks for
macros defined in a header file before. In this case, they should
not be `#undef`ed at the end of the file. This was an issue for
the AVX2 constant table offsets, which broke the native monobuild.

This commit fixes `autogen` by not emitting `#undef`s for macros
which are defined in a header as well. In that case, a comment is
emitted instead.

The check is not perfect because we don't check that the C file does
actually include the header with the overlapping macro, but it seems
good enough.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>