autogen: Allow importing 'clean' AArch64 backend
With `autogen --aarch64-clean`, the 'clean' AArch64 backend will be
imported from `dev/aarch64_clean/*` into `mlkem/native/aarch64/*`.

With `autogen --no-simplify` -- which can be combined with `--aarch64-clean`,
simplification can be skipped.

Care has to be taken to not use `git ls-files` to list source files,
since that's only updated once files are checked in. That does not
work for `autogen`, which needs to create files and modify them afterwards
(e.g. when copying from `dev/*` to `mlkem/*` and then adjusting headers),
or when deleting files and then of course not modifying them anymore.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>