CI: Only cache one valgrind per arch
We are using a patched valgrind for the ct tests
which is built in CI and  then cached in the Github Cache.
Currently we are using many different compilers for these
tests. Currently, for each of these compilers valgrind
is recompiled and cached taking up 600-800 MB per
each of the 26 compiler/arch combinations.
We only have 10 GB of cache, so we are constantly triggering
cache evictions.
This leads to painfully slow tests for other tests that due to
that have to recompile gcc which takes around 45 minutes.

This commit changes the cache key for valgrind to only contain
the OS and arch, but not the compiler version.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>