CI: Add dedicated job for running autogen
The `lint` job includes `autogen`, which includes the simplification
of assembly. Since we have assembly files for multiple architectures,
however, full assembly simplification requires a cross toolchain.

To avoid requiring the full-blown `ci-cross` shell for the `lint` job,
this commit changes `autogen` so that it skips over cross-simplification
unless the `--force-cross` parameter is given.
The `lint` job is then called without cross simplification, but instead
run on both an aarch64 and x86_64 machine, which jointly cover the entire
assembly.

As an additional safeguard, and also to check that simplification via
cross-toolchains work, a separate CI job for the checking of autogenerated
files is introduced which does run in the ci-cross shell and simplifies
all assembly.

The previous `simpasm` CI job is no longer needed (since the simplified
asm is now part of the source tree) and removed. Instead, a job is added
which reinstates the clean/opt backend with/without simplification, and
tests it.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>