autogen: Adjust header guards at time of update-via-copy
Previously, `autogen` would update the backend headers in `mlkem/*`
in a two-step fashion: First, synchronize them with the headers in `dev/*`,
and then update the guards. This approach works in a plain `autogen`, since
_overall_, the file in `mlkem/*` should not change if it was already up to
date. However, in a `autogen --dry-run` run, it will fail in the interim
stage where the header from `dev/*` is copied but the header guard is not
yet adjusted.

This commit fixes this by adjusting the header guard as part of the
synchronization from `dev/*` to `mlkem/*`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>