Benchmark: Add new benchmarking shell
https://github.com/pq-code-package/mlkem-native/pull/747 switched
the benchmarking from the ci shell to the ci-gcc14 shell.
While this is sufficient for most platforms, it breaks benchmarks on
platforms that do not run the Github actions locally, but instead
need binaries copied over the network and run through ssh.

This PR fixes that as well and adds a new dedicated benchmarking shell
which is the same as the ci shell prior to #740.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>