autogen: Manually handle #undef's for cbmc.h
cbmc.h defines various assertions only if CBMC is set, which
will not be the case for a production build. #undef'ing all
those macros in the monobuild file is not only unnecessary,
but also risky, since their names are very generic and may
well be defined elsewhere. Then, #undef'ing them would break
the build.

This commit excludes cbmc.h from the automatic generation of
`#undef` directives, and #undef's explicitly what is defined
even when CBMC is unset.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>