Make custom FIPS-202 headers configurable
While consumers of mlkem-native can bring their own FIPS-202 implementation
(as demonstrated by libOQS and the example/bring_your_own_fips202) it is
slightly rigid in that the FIPS-202 headers have to be reachable through a
path "fips202/fips202.h" and "fips202/fips202x4.h". In the case of libOQS,
this is the reason why a patch is needed.

This commit makes the installation of a custom FIPS-202 implementation more
convenient, by adding the configuration options
- MLK_FIPS202_CUSTOM_HEADER
- MLK_FIPS202X4_CUSTOM_HEADER
If set, those must be the names of files replacing fips202/fips202.h and
fips202/fips202x4.h If they are not set (default), mlkem-native's own
FIPS-202 implementation will be used.

This is demonstrated in the case of example/bring_your_own_fips202, where
we can now freely rename the directory holding the tiny_sha3 implementation.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>