autogen/monobuild: Guard inclusion of FIPS-202 sources
This commit makes the monobuild file more configurable by guarding
the inclusion of fips202 files, as well as the #undef'ing of the macros
they define, by !defined(MLK_MONOBUILD_CUSTOM_FIPS202).

This is needed for monobuilds with a custom FIPS-202, AWS-LC being
one example.

A comment is added that the `unifdef` tool is useful for statically
simplifying the monobuild file for reduced configurations.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>