Inline polyvec_basemul_acc_montgomery
This commit removes the function polyvec_basemul_acc_montgomery
and inlines its body at its single call site in `indcpa_dec()`.

The main benefit of this is a that the temporary buffer used for
the mulcache is now at the top of `indcpa_dec()` and can be zeroized
alongside the other stack buffers.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>