Destruct intermediate buffers on the stack
FIPS-203, Section 3.3 demands the destruction of intermediate values
before returning to the caller. This commit implements this.

A new function `ct_zeroize()` is introduced which is used to wipe
intermediate stack buffers prior to function return.
(NB: The constant-time'ness is not relevant here, but
conceptually the function is still best-placed in verify.h alongside
other functions which, when naively implemented, would be prone to
harmful compiler optimization -- ct_zeroize fits this category).

By default, ct_zeroize() is implemented via SecureZeroMemory on
Windows, and a plain memset + compiler barrier otherwise. If neither
makes sense, compilation fails.
Using memset_s would be preferred, but there is no portable way of
detecting whether it is available.

If users need to register a custom ct_zeroize, they can do so by
defining MLK_CT_ZEROIZE_NATIVE and defining ct_zeroize_native,
similar to how the arithmetic and FIPS-202 backends work.
This includes the case where the use does not wish to do stack
zeroization, or has an entirely separate (and more robust) method
for it, such as compiler instrumentation. In this case,
`ct_zeroize_native()` can be set to a no-op.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>