Add Mingw-w64 quickcheck
This commit adds a quickcheck using the Mingw-w64 compiler on Windows.
We install Mingw-w64 using Chocolatey which comes preinstalled on the
Github Windows runners.

To make the testvector hashes match, we need to set stdout to binary mode as
otherwise \n gets automatically converted to \r\n causing the hash to not match.
We add this behind preprocessor conditionals in gen_KAT.c and gen_NISTKAT.c.

Note that our Makefile heavily depends on having a Linux-like shell for commands
such as tr. We work around this by running the quickcheck in a bash shell
(which comes with Git for Windows which is preinstalled on the runners).

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>