Explicitly mark external API with MLK_EXTERNAL_API
This commit adds MLK_EXTERNAL_API annotations to the toplevel API
of mlkem-native. By default, MLK_EXTERNAL_API is defined as the
empty string, hence giving external linkage to the public API.

However, in a monobuild where both mlkem-native and the consuming
application are in the same compilation unit, MLK_EXTERNAL_API can
be defined as `static`, hence giving internal linkage to mlkem-native.

This is exemplified through modification of the example
`monolithic_build_multilevel_native`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>