Use random coins for Pairwise Consistency Test (PCT)
Previously, the PCT used static coins. However, FIPS 203 explicitly
says that PCT coins need to be randomly chosen. This commit adjusts
the PCT accordingly.

The reason why the coins were statically chosen in the first place
was to not alter the PRNG state and thereby invalidate NIST-KAT tests.
With random coins in the PCT, we have to remove PCT + NIST KAT tests
from the CI. This is done by disabling the PCT in most tests, but adding
a dedicated 'custom config' test for the PCT, which runs functionality,
KAT and ACVP tests only. PCT remains enabled in the CBMC proofs.

The PCT is also enabled in the CT tests, which wasn't previously the case.
In fact, the CT tests previously failed with PCT, for multiple reasons:
- First, since the PCT runs the PK check and thus requires the PK to be
  declassified -- however, PK declassification was previously done only
  after the PCT success. This commit changes this to declassify the PK
  before running the PCT.
- Second, since the _result_ of the PCT depends on secret data, but
  is used in a branch determining the return value of the key generation.
  This is fixed by declassifying the PCT result (but nothing else).

Finally, the KAT tests for all examples need to be disabled because
they rely on the PRNG state not being changed in between key generation
and encapsulation.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>