Add runtime option to break PCT
This commit introduces a config option MLK_KEYGEN_PCT_BREAKAGE_TEST.
When set, the user must provide a runtime (!) function mlk_break_pct()
to indicate if the PCT should be made fail. If set, the shared secret
in the PCT will be deliberately corrupted to make the PCT fail.
A test is added to exercise the runtime breakage logic.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>