CBMC: Fix include path
When CBMC, debug assertions are transformed into proof obligations
in debug.h. For this, debug.h includes cbmc.h, but using the wrong
include path "../cbmc.h" which is still a remnant of debug.[ch] living
in a separate debug/ directory. This only works because fips202/
happens to be another directory in the include path.

This commit fixes this minor slip by including "cbmc.h" from debug.h
directly.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>