Normalize naming of NTT "Layers"
The C front-end and back-ends and other documentation consistently
refer to the "layers" of the NTT as "1" through "7".

Previously, the assembly language back-end for AArch64 referred
to the layers as "0" through "6".

This commit changes the AArch64 back-end to use "1" through "7"
for consistency with the C code throughout.

Signed-off-by: Rod Chapman <rodchap@amazon.com>

Update optimized versions of AArch64 NTT back-end to used
revised names for NTT Layers.

Signed-off-by: Rod Chapman <rodchap@amazon.com>