CBMC: Add script to check that contracts have proofs
This commit adds a script `scripts/check-contracts` which looks for
functions with `__contract__` annotations, and checks that a proof
for them has been given -- with a few stated exceptions.

The script is added to CI.

The approach is not perfect since we do not consider which compile-time
options may influence the behaviour of a function -- but it's a start.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>