Poly: Hardcode barrett multiplier
In most places in the code, we use magic constants with
`check-magic: ...` annotations explaining their origin,
rather than computing the magic constant in the code
and relying on constant evaluation by the compiler.

One place where this was not done is `barrett_reduce()`,
where the 'magic' Barrett multiplier was computed in C.

This commit aligns `barrett_reduce()` to other routines
relying on magic constants, by hardcoding the magic value
in C, and adding a `check-magic: ...` annotation to explain
its origin.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>