HOL-Light: Add correctness proof for poly_tomont
This commit adds a HOL-Light functional correctness proof for
the AArch64-native version of `poly_tomont`.

The proof is structurally similar to that for the NTT and re-uses
core proof tools, esp. the lemma about correctness and bound of
Barrett multiplications.

The proof executes in a few minutes on a modern machine, but
it is notable how some proof steps are significantly slower
than others -- similar to the NTT; the proof may be a good
test ground for performance improvements.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>