HOL-Light: Add functional Keccak spec
This commit imports the functional spec for Keccak in HOL into
the mlkem-native source tree. Previously, those were indirectly
imported from s2n-bignum as part of the runtime nix setup, preventing
direct inspection of the assembly specs from the source tree.

The HOL-Light Github workflow is refined slightly to allow for
the specification of per-proof dependencies.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>