CBMC: switch --native-compiler to $(CC) instead of gcc
Currently CBMC defaults to use gcc. This is a problem when on MacOS, where
we default to clang. On these systems our CBMC proofs only work if the user
happens to have gcc in their $PATH.
This commit changes the flags for goto-cc to use the compiler set in $CC which
is clang on MacOS and gcc on other platforms.

See https://github.com/pq-code-package/mldsa-native/pull/65
Thanks https://github.com/jakemas for reporting this issue!

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>