HOL-Light: Add low-level spec and proof for AArch64 basemul
This commit adds low-level specifications and functional correctness proofs
for the AArch64 polynomial base multiplication routines

- poly_basemul_acc_montgomery_cached_k2
- poly_basemul_acc_montgomery_cached_k3
- poly_basemul_acc_montgomery_cached_k4

The specs are 'low-level' in the sense that they do only provide a
description in HOL of what is being computed on raw words; they do,
however, not yet give a high-level description of what the computation
does mod-3329: A scalar product in `Prod_i F_q[X]/(X^2-zeta_i)`.

The addition of a high-level spec and the connection of low-level and
high level spec will be in a future commit.

s2n-bignum is updated to incorporate the recent addition of various
long multiplication instructions.

The other ML-KEM arithmetic proofs needed updating since a core
tactic SIMD_SIMPLIFY_TAC for the simplification of SIMD expressions
was generalized to allow for a parametrized list of definitions to
fold. Previously, this was hardcoded to the definitions used by the
NTT and invNTT.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>