Define `mlk_polyvec[_mulcache]` as typedef for array
This commit changes the definition of `mlk_polyvec` from a struct
wrapper `struct { mlk_poly vec[MLKEM_K] }` to a typedef for
`mlk_poly[MLKEM_K]`. Similarly, it introduces `mlk_polymat`
as a typedef for `mlk_poly[MLKEM_K * MLKEM_K]`.

This ensures that when working with polynomial matrices
-- that is, arrays of `mlk_polyvec`s -- the polynomial
entries are consecutive in memory. This needs to be assumed
for the correctness of `gen_matrix()`.

From all we know, is a theoretical concern that without this
change, compilers would be free to introduce padding into
`mlk_polyvec`, breaking `gen_matrix()`. Yet, the behaviour of
`gen_matrix` is strictly speaking UB, so even if a compiler
does not use padding in `mlk_polyvec`, it is better to change it.

Defining `mlk_polymat` as a typedef for `mlk_poly[MLKEM_K * MLKEM_K]`
and not `mlk_polyvec[MLKEM_K]` is deliberate: In the latter case, one
would not be able to pass a `mlk_polymat` to a function expecting
`const polymat`, forcing cumbersome manual casts.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>