Define `mlk_polyvec[_mulcache]` as typedef for array
This commit changes the definition of `mlk_polyvec` from a struct
wrapper `struct { mlk_poly vec[MLKEM_K] }` to a typedef for
`mlk_poly[MLKEM_K]`.

This ensures that when working with polynomial matrices
-- that is, arrays of `mlk_polyvec`s -- the polynomial
entries are consecutive in memory. This needs to be assumed
for the correctness of `gen_matrix()`.

From all we know, is a theoretical concern that without this
change, compilers would be free to introduce padding into
`mlk_polyvec`, breaking `gen_matrix()`. Yet, the behaviour of
`gen_matrix` is strictly speaking UB, so it is better to change it.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>