Add SLOTHY nix setup
This installs SLOTHY in the default shell and the ci shell through nix.
This makes slothy-cli available in the PATH.

Unfortunately, the packaged ortools in nixpkgs does not find solutions in
reasonable time even though it is the same release (v9.7) as the one we are
using in the SLOTHY pip setup. I instead wrote a flake that fetches ortools
v9.7 from pypi.
Newer versions (including the most recent v9.12) unfortunately result in much
poorer performance.

Another issue I had to work around is that the unicorn version (2.1.1)
available in nixpkgs does not work on MacOS. I fetch the more recent 2.1.3
from pypi instead - this can be changed once nixpkgs is updated.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>