HOL-Light: Autogenerate zeta tables in proof scripts
This commit extends `autogen` to autogenerate the OCaml list
of zeta values used in the HOL-Light specs for the AArch64
forward and inverse NTTs.

This is reusing the same generator for the zeta values that is
already used for the autogeneration of the C zeta tables, but
uses a new formatter for OCaml.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>