SLOTHY: Update ortools to 9.12
This updates ortools used in SLOTHY to 9.12. In our CI we significantly better
performance (5.5-6 minutes vs. 8-10 minutes for the optimization of all asm).

ortools v9.12 requires a newer protobuf version than what is in nixpkgs - we
instead fetch it from pypi directly.
There is quite a bit of potential in this PR for clean-up when nixpkgs 25.05
is released.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>