Update LLVM/clang to 16 for MacOS and 18 for llvmPackages
nixpkgs 24.11 updated the default llvm/clang from 15->16 for MacOS and 17->18
for other platforms. We did not adopt that update resulting in multiple
llvm versions being installed on those systems which is unnecessary.
This commit aligns our versions with the default versions in nixpkgs.

With the upcoming 25.05 both MacOS and other platforms will move to
llvm/clang 19 - then we can simply drop the version numbers.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>