CI: Disable namespace checking when asan is enabled
With the update to clang16, we see some CI failures on MacOS when our
namespace checking script is used in combination with address sanitizer as
asan inserts some symbols which are then not appropriately namespaced.

This commit adds a configuration option to functest and multi-functests
that allows disabling the namespacing check.
This check is then disabled for the offending workflows where asan is used.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>