CI: Implement simple retry-once logic for starting EC2 runners
We often need to manually restart the CI because too many EC2 instances
are attempted to be started at the same time.

This commit implements a basic retry-once logic in ci_ec2_reusable.yml:
Upon failure to start an EC2 instance, it waits ~30s and retries once
before failing. A slight random delay is inserted to reduce the likelihood
of clashes if multiple jobs fail at the same time.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>