Fix typo in mlkem/mlkem_native.h and test in multi-level builds
**Context**

mlkem_native.h defines the public API exposed by mlkem-native.
It can be used in two forms:

- Standalone

  The config.h used to build mlkem-native is not present.
  Instead, the user sets the level and namespace explicitly before
  including mlkem_native.h

- Config-based

  The original config.h is present, and used by mlkem_native.h to
  derive namespace and level.

mlkem_native.h also has to consider multi-level builds, where the
level should be added to the namespace prefix.

**Issue**

There was a typo in the config-based multi-level version of mlkem_native.h.
This typo went unnoticed because this usage of mlkem_native.h was never tested.

This commit fixes the typo and adds a test exercising mlkem_native.h
in the config-based multi-level version. Specifically, the existing
monolithic_build_multilevel example is adjusted to use mlkem_native.h
_both_ in config-based and standalone forms now, ensuring consistency
of the two.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>