Autogen: Autogenerate comments for `#else` and `#endif`
In parts of the mlkem-native source, we annotate `#else`
and `#endif` directives with the condition to which they apply.
This can be useful, but is also tedious to do by hand, and difficult
to do consistently.

This commit extends `autogen` with the ability to autogenerate
the `#else` and `#endif` annotations.

It was found that appyling comments everywhere impedes readability
when `#if`, `#else` and `#endif` are very close to each other and
there is no need to remind the reader of what condition is being
closed. We therefore only apply comments if the distance between
`#if` and `#else` and `#endif` is larger than a certain threshold
(for now, it's 5), or if there are nested `#if`s inside.

A very basic parser and simplifier for conditionals is implemented
based on `pyparsing`, which takes care of removing unnecessary brackets,
normalizing spacing, and removing double-negations.

Further, `autogen` is taking >10s even on a powerful machine. This commit
adds some status logging to indicate progress and where the time is spent.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>