autogen: Autogenerate #else and #endif comments for .S files
A previous commit extended `autogen` to autogenerate comments
for `#else` and `#endif` directives, indicating the conditions
they guard.

This commit extends this feature to .S files which were previously
excluded because we cannot run `clang-format` on them. This is
circumvented here by running `clang-format` on individual lines
with preprocessor directives.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>