Config: Use `MLK_CONFIG_XXX` naming scheme for configuration options
This commit renames all configuration options to follow the pattern
`MLK_CONFIG_XXX`.

One exception is MLKEM_K, which stays as-is in the bulk of the code.
However, instead of exposing MLKEM_K directly in the config, it is
now determined by the value of a new configuration option
MLK_CONFIG_SECURITY_LEVEL, which takes the value 512/768/1024.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>