Simplify mlkem_native.h
The goal of mlkem/mlkem_native.h is to capture the public API of
mlkem-native in a way that is independent of the main source tree.

Previously, mlkem_native.h could be used in two ways:
- Standalone: Here, prior to inclusion of mlkem_native.h, the user
  would set various MLK_BUILD_INFO_XXX flags indicating parameter set
  and namespacing.
- Config-based: Here, the user would provide the configuration used
  to build mlkem-native, and have MLK_BUILD_INFO_XXX derived
  automatically.

The second case is mainly for convenience during development: When
the mlkem-native source code is available, one may as well include
kem.h and use it as the API. We include mlkem_native.h in this case
primarily to test consistency of mlkem_native.h and kem.h.

This commit aims to simplify the use of mlkem_native.h and to align
it with other configuration of mlkem-native, as follows:

First, MLK_BUILD_INFO_XXX are renamed to follow the naming pattern
MLK_CONFIG_API_XXX. In particular, this is a sub-pattern of the
pattern MLK_CONFIG_XXX used for options that the user may tweak --
which is the case here.

The options controlling the use of mlkem_native.h, are then:

- MLK_CONFIG_API_PARAMETER_SET [required]

  The parameter set: 512, 768, 1024

  Corresponds to MLK_CONFIG_PARAMETER_SET used in the config.h.

- MLK_CONFIG_API_NAMESPACE_PREFIX [required]

  The namespace prefix, including level.

  This almost corresponds to MLK_CONFIG_NAMESPACE_PREFIX used
  during the build, but includes the level.

- MLK_CONFIG_API_NO_SUPERCOP [optional]

  This used to be MLK_BUILD_INFO_NO_STANDARD_API, but the
  new name seems more apt.

- MLK_CONFIG_API_CONSTANTS_ONLY [optional]

  This can be set if the header is included only to be the
  sizes of key material. This is needed in monolithic multilevel
  builds.

Second, the config-based usage of mlkem_native.h is removed. The
only way to use mlkem_native.h is to set a subset of the above
MLK_CONFIG_API_XXX options and then include mlkem_native.h.

To not loose the convenience of importing mlkem_native.h during
a) an mlkem-native build for consistency checking, or b) a test
application, the standard header mlkem/common.h is adjusted to
set MLK_CONFIG_API_XXX consistent with MLK_CONFIG_XXX, so that
one can just import mlkem_native.h whenever needed in any context
that has included mlkem/common.h.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>