keypair_derand: *coins -> coins[2 * MLKEM_SYMBYTES]
keypair requires 2 * MLKEM_SYMBYTES bytes of randomness. So far the function
signature took a const uint8_t *coins.
This commit changfes the function signature to make the size explicit:
const uint8_t coins[2 * MLKEM_SYMBYTES].
This is in line with the enc_derand function and should reduce the risk of
misuse.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>