tests: Integrate namespace check into `tests func`
This commit adds the parameter --check-namespace to `tests func`.
It is false by default. If true, `scripts/check-namespace` is called
after both the C and the native build to check namespacing.

Previously, a call to `check-namespace` after `tests func` would only
apply to the final binary produced by `tests func`, which was the native
one.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>