CI: Remove 'oldpqclean' from libOQS integration test
The libOQS integration tests fail because of build failures
related to SHA3 assembly in old PQClean implementations of
Kyber and Dilithium.

Checking the build for those old implementations is not the
goal of the mlkem-native->libOQS integration test.

This commit disables the offending implementations
by removing them from the YML file describing the
3rd party implementations imported into libOQS.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>