Document MLK_NAMESPACE and MLK_NAMESPACE_K macros
For readers trying to see through the naming indirection using
MLK_NAMESPACE(...) and MLK_NAMESPACE_K(...), the previous definitions

```c
 #define MLK_NAMESPACE(s) \
   MLK_CONCAT(MLK_CONCAT(MLK_CONFIG_NAMESPACE_PREFIX, _), s)
 #define MLK_NAMESPACE_K(s) \
   MLK_CONCAT(MLK_CONCAT(MLK_ADD_PARAM_SET(MLK_CONFIG_NAMESPACE_PREFIX), _), s)
```

are a handful.

This commit keeps the functionality as-is, but introduces internal
abbreviations for the level'ed and underscore'ed prefix, so that the
toplevel definitions for MLK_NAMESPACE[_K] turn into the more expectable

```c
 #define MLK_NAMESPACE(s) MLK_CONCAT(MLK_NAMESPACE_PREFIX, s)
 #define MLK_NAMESPACE_K(s) MLK_CONCAT(MLK_NAMESPACE_PREFIX_K, s)
```

Also, a comment is added with an example.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>