CI: Add compiler tests for zig
This commit adds compiler tests for zig to the CI.

Notably, because of an issue in the propagation of CPU
architecture CFLAGS to the assembler, we cannot build
native code with zig:
- https://github.com/ziglang/zig/issues/23576

Since the `[multi-]functest` action does not allow
for the selection of individual examples to be run,
 and some of them include native code, we also
 cannot test the examples with zig.

Finally, there is a build issue for zig-0.10 on Mac,
hence the respective test is disabled.

Still, we test the C-backend with zig versions 0.9-0.13
and various compilation options.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>