Makefile: Remove duplicated symbol check
test/mk/rules.mk includes a check for duplicated symbols when
building archive files.

The implementation of this check is specific to the toolchain:
We already distinguish Darwin vs. non-Darwin, and exclude the
check when running with scan-build.

It turns out that the check does also not work with zig, as
it uses unsupported linker arguments.

Instead of attempting to make the check more portable, this
commit removes it altogether. This is OK because we anyway
build various test binaries based on the generated libraries,
and the build for those binaries would fail if there were
duplicated symbols.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>